refactor: rename issues.py to issue_handler.py to avoid conflict with issues/ cmd dir
